Module restructuring to better delineate between safe/unsafe
