+++ src/Database/PostgreSQL/Tx/Unsafe.hs
@@ -1,0 +1,12 @@
+module Database.PostgreSQL.Tx.Unsafe
+  ( unsafeRunTxM
+
+  , UnsafeTx(unsafeIOTx)
+  , UnsafeUnliftTx(unsafeWithRunInIOTx)
+
+  , unsafeRunIOInTxM
+  , unsafeRunTxInIO
+  , unsafeReaderIOTx
+  ) where
+
+import Database.PostgreSQL.Tx.Internal
